perf(gc): split arming into its own module; fmt
